seed database, create middleware

+++ Core/Helpers/JwtOptions.cs
@@ -4,6 +4,6 @@ public class JwtOptions
 {
     public string Issuer { get; set; }
     public string Key { get; set; }
-    public int LifeTime { get; set; }
+    public float LifeTime { get; set; }
     public int RefreshTokenLifeTimeInDays { get; set; }
 }
